refactor: update file names to match export
