+++ lib/ArraySlice.js
