+++ lib/KeyValuePair.js
