+++ lib/Namespace.js
@@ -4,7 +4,7 @@ const isNumber = require('lodash/isNumber');
 const isBoolean = require('lodash/isBoolean');
 const isObject = require('lodash/isObject');
 
-const JSONSerialiser = require('./serialisers/json');
+const JSONSerialiser = require('./serialisers/JSONSerialiser');
 const elements = require('./elements');
 
 /**
