+++ lib/ObjectSlice.js
@@ -1,5 +1,5 @@
 const negate = require('lodash/negate');
-const ArraySlice = require('./array-slice');
+const ArraySlice = require('./ArraySlice');
 
 class ObjectSlice extends ArraySlice {
   map(callback, thisArg) {
