+++ lib/elements.js
@@ -1,18 +1,18 @@
-const Element = require('./primitives/element');
-const NullElement = require('./primitives/null-element');
-const StringElement = require('./primitives/string-element');
-const NumberElement = require('./primitives/number-element');
-const BooleanElement = require('./primitives/boolean-element');
-const ArrayElement = require('./primitives/array-element');
-const MemberElement = require('./primitives/member-element');
-const ObjectElement = require('./primitives/object-element');
-const LinkElement = require('./elements/link-element');
-const RefElement = require('./elements/ref-element');
+const Element = require('./primitives/Element');
+const NullElement = require('./primitives/NullElement');
+const StringElement = require('./primitives/StringElement');
+const NumberElement = require('./primitives/NumberElement');
+const BooleanElement = require('./primitives/BooleanElement');
+const ArrayElement = require('./primitives/ArrayElement');
+const MemberElement = require('./primitives/MemberElement');
+const ObjectElement = require('./primitives/ObjectElement');
+const LinkElement = require('./elements/LinkElement');
+const RefElement = require('./elements/RefElement');
 
-const ArraySlice = require('./array-slice');
-const ObjectSlice = require('./object-slice');
+const ArraySlice = require('./ArraySlice');
+const ObjectSlice = require('./ObjectSlice');
 
-const KeyValuePair = require('./key-value-pair');
+const KeyValuePair = require('./KeyValuePair');
 
 /**
  * Refracts a JSON type to minim elements
