+++ lib/elements/LinkElement.js
@@ -1,4 +1,4 @@
-const Element = require('../primitives/element');
+const Element = require('../primitives/Element');
 
 /** Hyperlinking MAY be used to link to other resources, provide links to
  * instructions on how to process a given element (by way of a profile or
