+++ lib/elements/RefElement.js
@@ -1,4 +1,4 @@
-const Element = require('../primitives/element');
+const Element = require('../primitives/Element');
 
 /**
  * @class RefElement
