+++ lib/minim.js
@@ -1,4 +1,4 @@
-const Namespace = require('./namespace');
+const Namespace = require('./Namespace');
 const elements = require('./elements');
 
 // Direct access to the Namespace class
@@ -9,7 +9,7 @@ exports.namespace = function namespace(options) {
   return new Namespace(options);
 };
 
-exports.KeyValuePair = require('./key-value-pair');
+exports.KeyValuePair = require('./KeyValuePair');
 
 exports.ArraySlice = elements.ArraySlice;
 exports.ObjectSlice = elements.ObjectSlice;
@@ -27,5 +27,5 @@ exports.LinkElement = elements.LinkElement;
 
 exports.refract = elements.refract;
 
-exports.JSONSerialiser = require('./serialisers/json');
-exports.JSON06Serialiser = require('./serialisers/json-0.6');
+exports.JSONSerialiser = require('./serialisers/JSONSerialiser');
+exports.JSON06Serialiser = require('./serialisers/JSON06Serialiser');
