+++ lib/primitives/ArrayElement.js
@@ -1,6 +1,6 @@
 const negate = require('lodash/negate');
-const Element = require('./element');
-const ArraySlice = require('../array-slice');
+const Element = require('./Element');
+const ArraySlice = require('../ArraySlice');
 
 /**
  * @class
