+++ lib/primitives/Element.js
@@ -1,6 +1,6 @@
 const isEqual = require('lodash/isEqual');
-const KeyValuePair = require('../key-value-pair');
-const ArraySlice = require('../array-slice.js');
+const KeyValuePair = require('../KeyValuePair');
+const ArraySlice = require('../ArraySlice.js');
 
 /**
  * @class
