+++ lib/primitives/MemberElement.js
@@ -1,5 +1,5 @@
-const KeyValuePair = require('../key-value-pair');
-const Element = require('./element');
+const KeyValuePair = require('../KeyValuePair');
+const Element = require('./Element');
 
 /**
  * @class MemberElement
