+++ lib/primitives/NullElement.js
@@ -1,4 +1,4 @@
-const Element = require('./element');
+const Element = require('./Element');
 
 module.exports = class NullElement extends Element {
   constructor(content, meta, attributes) {
