+++ lib/primitives/NumberElement.js
@@ -1,4 +1,4 @@
-const Element = require('./element');
+const Element = require('./Element');
 
 /**
  * @class NumberElement
