+++ lib/primitives/ObjectElement.js
@@ -1,9 +1,9 @@
 const negate = require('lodash/negate');
 const isObject = require('lodash/isObject');
 
-const ArrayElement = require('./array-element');
-const MemberElement = require('./member-element');
-const ObjectSlice = require('../object-slice');
+const ArrayElement = require('./ArrayElement');
+const MemberElement = require('./MemberElement');
+const ObjectSlice = require('../ObjectSlice');
 
 /**
  * @class
