+++ lib/serialisers/JSON06Serialiser.js
@@ -1,6 +1,6 @@
-const JSONSerialiser = require('./json');
+const JSONSerialiser = require('./JSONSerialiser');
 
-module.exports = class JSONSerialiser06 extends JSONSerialiser {
+module.exports = class JSON06Serialiser extends JSONSerialiser {
   serialise(element) {
     if (!(element instanceof this.namespace.elements.Element)) {
       throw new TypeError(`Given element \`${element}\` is not an Element instance`);
