+++ lib/serialisers/JSONSerialiser.js
