+++ test/ObjectSlice-test.js
