+++ test/primitives/StringElement-test.js
